+++ core/src/sn/LayerMng.ts
@@ -40,8 +40,7 @@ export class LayerMng implements IGetFrm {
 	constructor(private readonly cfg: Config, private readonly hTag: IHTag, private readonly appPixi: Application, private readonly val: IVariable, private readonly main: IMain, private readonly scrItr: ScriptIterator, private readonly sys: SysBase, readonly sndMng: SoundMng, readonly alzTagArg: AnalyzeTagArg, readonly prpPrs: IPropParser) {
 		// レスポンシブや回転・全画面切り替え・DevTools 表示切り替えの対応
 		const fncResizeLay = ()=> {
-			if (! sys.cvsResize()) return;
-
+			sys.cvsResize();
 			this.cvsResizeDesign();
 			if (this.#modeLnSub) this.#aLayName.forEach(
 				layer=> this.#hPages[layer].fore.cvsResizeChildren()